requirements.txt
Django
Pillow
django-ckeditor
